Add Platform metrics processing
These changes allow for populating dedicated structs by using the Lambda platform report provided after each invocation. Some fields also need to be added to the Event structs.

A notable addition is the modification of the metadata to incorporate extension-specific fields. I think that this information should be included, but the current implementation causes the extension to be displayed as a separate service in the APM UI.