Support Lambda Platform Metrics (#202)
* Add model package for metrics

This commit is heavily inspired by the Go agent codebase. The data model needed to (un)marshal metrics as well as the associated metadata is defined.

* Add version utility file

This file contains information about the extension version. This used to enrich Metrics metadata with the version of the extension used to generate these metrics.

* Add metadata extraction from the first received transaction

This commit allows for the extraction of the metadata from the first received transaction, with the assumption that this metadata will persist across invocations between cold starts.

Metadata is obtained from the uncompressed payload, and a dedicated container is populated.

* Add Platform metrics processing

These changes allow for populating dedicated structs by using the Lambda platform report provided after each invocation. Some fields also need to be added to the Event structs.

A notable addition is the modification of the metadata to incorporate extension-specific fields. I think that this information should be included, but the current implementation causes the extension to be displayed as a separate service in the APM UI.

* Update logs API to handle Lambda platform metrics

* Update APM Transport to handle metadata population

* Update main to handle metadata and metrics

* Add end-to-end testing for metrics and metadata

* Fix: Populate event timestamp

* Fix: linting

* Fix: Implement review comments (unit tests)

* Fix: Use standard APM data model

* Fix: remove metadataJSON encoding / decoding

* E2E tests update

* Use lower_snake_case for metrics names

* Update process_metadata.go

* Small fixes (go mod, metrics names)

* Update notice

* Fix typo

* Fix : Use v2 APM model

* Resolve conflicts with main