fix: properly handle http.ErrServerClosed
Logs API server: the server was ignoring ErrServerClosed and logging an error.

APM Data server: the server was using an equality check with the error message but
it was never satisfied because of a typo ('server closed' vs 'Server closed'), leading
to an error message.

Both servers are now handling the error and logging a message that the server stopped
successfully. The code is using go 1.13 errors to avoid comparing strings and for better
compatibility.