fix: randomise the initial grace period to avoid collisions
The previous algorithm was using binary exponential-backoff with
a +- 10% jitter to calculate the grace period.
Because there can be multiple lambda environments we need to
mitigate collisions:

We cannot use 0 as the first delay because functions failing closer to
each other will collide. The issue would then be propagated by the
small jitter for lower delays.

This change adds an initial delay of n seconds to the first reconnection
attempt.
n is randomly generated in a closed interval to account for collisions
while keeping in mind usability and user experience.