refactor: redesign logs api
Use a logsapi client and add functional options to make it more
flexible in local or testing environments.

Add support for graceful shutdown to the HTTP server and only stop it
after the event has been processed to avoid race conditions on the log
channel.

Add more detailed errors and make sure they are wrapped to provide
appropriate context.

Rename and move code around for clarity.