refactor: remove deprecated ioutil references
io/ioutil was deprecated in Go 1.16.

See https://tip.golang.org/doc/go1.16#ioutil

This is needed to support newer version of golangci-lint because
the linters are flagging the usage of io/ioutil as an issue.
Because CI is using the linter for the lint job this is blocking
the update to newer version of go.