fix: avoid panic from closing the channel twice on warm starts
During multiple invocations the lambda can reuse the environment if
a warm start is taking place. We cannot assume a request with 'flushed=true'
will be the last one for the lifetime of the application.
Replace the channel with a counter that is increased when we
receive a request with 'flushed=true' and it is decreased if
we meet such request in the buffered data while sending to the APM
server.