fix: try to return as soon as possible on flush signal
The flush signal is received on a separate goroutine (http handler) so we
cannot assume anything on its relationship with the event processed by other
goroutines.
If we just check once we might miss the signal and hang until the runtimeDone
or timeout event is received.
To prevent this, create a channel and periodically check the flush counter to
minimize latency.