fix: always flush data to apm before shutting down and rework agent done signal (#258)
* fix: always flush data to apm before shutting down and rework agent done signal

Add a defer statement to make sure that we always flush data to the
apm server before shutting down.

Remove agent done signal channel and avoid leaking implementation details.
The channel was being recreated and closed on each event, racing with the
intake handler that was sending to the channel.
The channel is now used internally by the apm client and external packages
can call 'Done()' to check whether the agent has sent the final intake
request.

See https://github.com/elastic/apm/blob/main/specs/agents/tracing-instrumentation-aws-lambda.md#data-flushing

* sigh

* fix: avoid panic from closing the channel twice on warm starts

During multiple invocations the lambda can reuse the environment if
a warm start is taking place. We cannot assume a request with 'flushed=true'
will be the last one for the lifetime of the application.
Replace the channel with a counter that is increased when we
receive a request with 'flushed=true' and it is decreased if
we meet such request in the buffered data while sending to the APM
server.

* fix: try to return as soon as possible on flush signal

The flush signal is received on a separate goroutine (http handler) so we
cannot assume anything on its relationship with the event processed by other
goroutines.
If we just check once we might miss the signal and hang until the runtimeDone
or timeout event is received.
To prevent this, create a channel and periodically check the flush counter to
minimize latency.

* refactor: remove busy loop and rely on channels to signal flush requests

* fix: update behaviour based on flush strategy

only synchronously flush on sendstrategy == syncflush. Do not
flush just because there are unhandled flushed=true requests.

* fix: update flush logic and remove flush count

Go back to a less disruptive change. Remove flush count, don't
keep track of multiple flushed requests but reset the channel
before processing the event.

* refactor: move flush reset inside process event

* refactor: revert queue changes

Co-authored-by: Andrew Wilkins <axwalk@gmail.com>

* fix: move reset flush to a defer call

Co-authored-by: Andrew Wilkins <axwalk@gmail.com>