fix: guard status operations behind a mutex to avoid race conditions
Any change to the transport status is guarded by a mutex but
some methods were using the field directly, causing race
conditions.

apmproxy tests are using a timeout of 5s to wait for the status
to change during backoff. However there is a subtle bug that
causes the test to fail because the initial delay is exactly 5s.
To avoid this, the tests wait 7s now.

Update status name and documentation.