test: bump main test timeout to 20s
main tests were previously using 5s timeout but that leads to flaky tests
because the extension uses contexts with 5s delay to shutdown the logapi
and the receiver.

Timeouts are hardcoded and there's no way to change them as of now so the
immediate solution is to increase the test timeout.