build: make ess version configurable
The terraform script used for smoke testing was defaulting to the
latest version of ess.
Because we want to test on stable versions and allow people to reuse
the script to provision and test the dev environment, add a variable
to make ess version configurable and default to the latest 8.x version.