fix: do not return status 200 on server info request when unreachable
When proxying info requests to apm server we're loggin an error if
the server is unreachable.
The ErrorHandler should also write a status code back to avoid
returning a Status 200.
Add ModifyResponse to log a debug message with the body on unexpected
responses and let the ErrorHandler proxy the status code back to the
agent.
A StatusBadGateway is returned for generic errors.