docs: use the 'latest' GH release API for the latest extension or agent release
The "Monitoring AWS Lambda $lang Functions" doc pages use some
in-browser JS to interpolate the ARNs for the latest Lambda extension
release and the latest APM agent release. This changes that JS to use
the "latest release" GH API
    https://api.github.com/repos/elastic/${ghRepo}/releases/latest
rather than listing all (or the first page of GH releases):
    https://api.github.com/repos/elastic/${ghRepo}/releases
and picking the latest by `release.created_at`.

GH releases for a repo have a sense of the "latest" release
(see `--latest` flag to https://cli.github.com/manual/gh_release_create)
even if it isn't the latest tag or published release in time. The
Node.js APM agent will soon have 4.x releases and 3.x maintenance
releases. It is possible that a 3.x release is the latest in time, but
not marked as the "latest" release. This change ensures that the
Lambda doc page will pick the correct (4.x) one.